+++ pysui/__init__.py
@@ -41,7 +41,6 @@ SDK_CURRENT_VERSION: list[str] = [int(x) for x in __version__.split(".")]
 
 from pysui.sui.sui_clients.sync_client import SuiClient as SyncClient
 from pysui.sui.sui_clients.async_client import SuiClient as AsyncClient
-from pysui.sui.sui_clients.subscribe import SuiClient as Subscribe
 
 from pysui.sui.sui_common.config import PysuiConfiguration
 from pysui.sui.sui_pgql.pgql_clients import SuiGQLClient as SyncGqlClient