+++ pysui/sui/sui_pgql/pgql_configs.py
@@ -195,6 +195,6 @@ class SuiConfigBetaGQL:
 
 def pgql_config(env: str, sversion: Optional[str] = None) -> tuple[str, Callable]:
     """Get the configuration for Sui GraphQL."""
-    if env == "testbeta" or env == "mainbeta":
+    if env == "betatest" or env == "betamain":
         return _QUERY_BETA, SuiConfigBetaGQL.from_query
     return _QUERY, SuiConfigGQL.from_query